requirements.txt
psycopg2-binary
spacy==2.2.4
boto3
thinc
